Updated pom.xml and some dependencies Updated to 1.7.10 (Probably runs on other versions of Bukkit)

+++ src/main/java/syam/flaggame/listener/FGPlayerListener.java
@@ -533,7 +533,7 @@ public class FGPlayerListener implements Listener {
                 }
 
                 // 20以上にならないように体力とお腹ゲージを+2(ハート、おにく1つ分)回復させる
-                int nowHP = player.getHealth();
+                double nowHP = player.getHealth();
                 nowHP = nowHP + 2;
                 if (nowHP > 20) nowHP = 20;
